+++ src/misc/cache/_lv_cache_lru_rb.c
@@ -458,5 +458,5 @@ static uint32_t cnt_get_data_size_cb(const void * data)
 static uint32_t size_get_data_size_cb(const void * data)
 {
     lv_cache_slot_size_t * slot = (lv_cache_slot_size_t *)data;
-    return slot->size;
+    return (uint32_t)slot->size;
 }
